+++ BattleNetwork/bnVirusBackground.cpp
@@ -19,7 +19,7 @@ VirusBackground::~VirusBackground(void) {
 }
 
 void VirusBackground::Update(float _elapsed) {
-  progress += 0.5 * _elapsed;
+  progress += 0.5f * _elapsed;
   if (progress >= 1.f) progress = 0.0f;
 
   x += 1 * _elapsed;
@@ -27,9 +27,9 @@ void VirusBackground::Update(float _elapsed) {
   if (x > 1) x = 0;
   if (y > 1) y = 0;
 
-  int frame = progress * COMPONENT_FRAME_COUNT;
+  int frame = (int)(progress * COMPONENT_FRAME_COUNT);
 
   Wrap(sf::Vector2f(x, y));
-  TextureOffset(sf::Vector2f(frame*COMPONENT_WIDTH, 0));
+  TextureOffset(sf::Vector2f((float)(frame*COMPONENT_WIDTH), 0));
 
 }
